refactor(api): organize src/ into scalable feature modules
- Split src/ into feature modules with clear boundaries
- Move shared utilities, constants, and enums to common module
- Improve folder structure for maintainability
- No breaking changes to existing functionality

Closes #476
