+++ meridian-api/src/metaoption/metaoption.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { MetaOption } from './metaoption.entity';
-import { CreatePostMetaOptionsDto } from 'src/DTO/createPostMetadto';
+import { CreatePostMetaOptionsDto } from './dto/create-post-meta-options.dto';
 
 @Injectable()
 export class MetaoptionService {
