+++ meridian-api/src/post/dto/post-param.dto.ts
@@ -1,0 +1,14 @@
+import { IsOptional, IsInt } from 'class-validator';
+import { Type } from 'class-transformer';
+import { ApiPropertyOptional } from '@nestjs/swagger';
+
+export class GetPostsParamDto {
+  @ApiPropertyOptional({
+    description: 'Get posts with a specific ID',
+    example: 1234,
+  })
+  @IsOptional()
+  @IsInt()
+  @Type(() => Number)
+  id?: number;
+}
