+++ meridian-api/src/tag/tags.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Tag } from './tag.entity';
 import { In, Repository } from 'typeorm';
-import { CreateTagDto } from 'src/DTO/createTagDto';
+import { CreateTagDto } from './dto/create-tag.dto';
 
 @Injectable()
 export class TagsService {
