+++ meridian-api/src/users/dto/user-param.dto.ts
@@ -2,11 +2,9 @@ import { IsOptional, IsInt } from 'class-validator';
 import { Type } from 'class-transformer';
 import { ApiPropertyOptional } from '@nestjs/swagger';
 
-// this was created because of the undefined in Param controller file
-// it is use to transfrom the id to number
-export class GetPostsParamDto {
+export class GetuserParamDto {
   @ApiPropertyOptional({
-    description: 'Get posts with a specific ID',
+    description: 'Get user with a specific ID',
     example: 1234,
   })
   @IsOptional()
